Move ast/astExtractor/types.ts to ast/types/ast.ts and BaseNode to ast/types/ast.ts
